flytter skoleår til egen fil
