+++ src/main/kotlin/no/nav/familie/ef/sak/felles/util/DatoUtil.kt
@@ -40,33 +40,3 @@ fun LocalDate.harPåfølgendeMåned(påfølgende: LocalDate): Boolean =
 fun YearMonth.erPåfølgende(påfølgende: YearMonth): Boolean = this.plusMonths(1) == påfølgende
 
 fun YearMonth.erSammeMåned(sammeMåned: YearMonth): Boolean = this == sammeMåned
-
-fun skoleår(fra: YearMonth, til: YearMonth): Skoleår {
-    brukerfeilHvis(til < fra) {
-        "Tildato=$til må være etter eller lik fradato=$fra"
-    }
-    if (fra.month > Month.JUNE) {
-        brukerfeilHvis(til.year == fra.year + 1 && til.month > Month.AUGUST) {
-            "Når tildato er i neste år, så må måneden være før september"
-        }
-        brukerfeilHvis(til.year > fra.year + 1) {
-            "Fradato og tildato må være i det samme skoleåret"
-        }
-        return Skoleår(Year.of(fra.year))
-    } else {
-        brukerfeilHvis(til.year != fra.year) {
-            "Fradato før juli må ha tildato i det samme året"
-        }
-        brukerfeilHvis(til.month > Month.AUGUST) {
-            "Fradato før juli må ha sluttmåned før september"
-        }
-        return Skoleår(Year.of(fra.year - 1))
-    }
-}
-
-data class Skoleår(val år: Year) {
-
-    override fun toString(): String {
-        return String.format("%ty/%ty", år, år.plusYears(1))
-    }
-}
